Move similar query logic to the separate module
