+++ crates/indexer/src/api/offers/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod db;
 pub(crate) mod dto;
 pub(crate) mod handlers;
+pub(crate) mod list_query;
 mod routes;
 
 pub use routes::routes;
