--- crates/indexer/src/api/participants/mod.rs
@@ -1,3 +1,0 @@
-mod scope;
-
-pub use scope::push_latest_participant_offers_scope;
